+++ src/main/java/co/edu/uceva/pokemon/persistence/entities/TiposEntity.java
@@ -11,9 +11,10 @@ public class TiposEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_tipo") // Asegúrate de que el nombre sea correcto
-    private Long idTipo; // Cambiado a Integer para permitir null
+    private Integer idTipo; // Cambiado a Integer para que coincida con DebilidadesId
 
     @Column(name = "nombre_tipo") // Opcional, pero buena práctica
     private String nombreTipo;
 
 }
+
